+++ src/app/(root)/(course)/course/[id]/admin/table/_table/declare.ts
@@ -1,4 +1,4 @@
-import { RowData } from "@tanstack/react-table";
+import { type RowData } from "@tanstack/react-table";
 
 declare module "@tanstack/react-table" {
   interface TableMeta<TData extends RowData> {
