+++ src/app/(root)/(course)/course/[id]/admin/table/_table/types.tsx
@@ -1,4 +1,4 @@
-import { RouterOutputs } from "@/trpc/react";
+import { type RouterOutputs } from "@/trpc/react";
 
 export type Unit = RouterOutputs["units"]["getTableData"][0];
